Refactor: RestTemlate 공유를 위한 새로운 패키지
