+++ investement/src/main/java/com/example/investment/index/service/client/IndexApiService.java
@@ -1,6 +1,6 @@
 package com.example.investment.index.service.client;
 
-import com.example.investment.news.service.client.RestTemplateClient;
+import com.example.investment.template.RestTemplateClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
